[Integration test] Yoast_Notification_Center: bugfix for partially supported callback
As things were, the callback which was being created would have the following form:
```
["Yoast_Notification_Center", "Yoast_Notification_Center::add_notification"]
```

Take note of the class name also being prepended to the method name.

This is due to the use of the `__METHOD__` magic constant, which includes the class name already. By changing this to `__FUNCTION__`, the method name _without_ the class name is returned.

As this type of callback is only partially supported by PHP, will be deprecated in PHP 8.2, with support being completely removed in PHP 9.0, fixing this seems prudent.

Refs:
* https://wiki.php.net/rfc/deprecate_partially_supported_callables